# openapi.yml
openapi: 3.0.3
info:
  title: ASPB Autowebinar API
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/v1/moderation/sessions:
    get:
      summary: List recent WebinarSessions available to the active tenant moderator
      security:
        - userSessionCookie: []
      parameters:
        - name: webinarId
          in: query
          schema: { type: string, minLength: 1, maxLength: 191 }
      responses:
        '200':
          description: Tenant-scoped sessions with real message, hidden-message and blocked-registration counts
        '403':
          description: Active membership is not OWNER or MODERATOR
        '404':
          description: Moderation dashboard is disabled
  /api/v1/moderation/sessions/{sessionId}/messages:
    get:
      summary: Read the exact tenant-scoped moderation projection for one WebinarSession
      security:
        - userSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/WebinarSessionId'
      responses:
        '200':
          description: Messages include exact type, synthetic flag, current visibility revision and registration chat state
        '403':
          description: Active membership is not OWNER or MODERATOR
        '404':
          description: Unknown and cross-tenant sessions are intentionally indistinguishable
  /api/v1/moderation/sessions/{sessionId}/messages/{messageId}:
    patch:
      summary: Hide or restore one message with optimistic revision and a mandatory reason
      security:
        - userSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/WebinarSessionId'
        - $ref: '#/components/parameters/ChatMessageId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatMessageModerationCommand' }
      responses:
        '200':
          description: Visibility changed immediately, room cache invalidated and audit written
        '400':
          description: Strict payload or reason validation failed; tenant fields are rejected
        '403':
          description: Active membership is not OWNER or MODERATOR
        '404':
          description: Unknown and cross-tenant messages are intentionally indistinguishable
        '409':
          description: Revision conflict or requested visibility is already current
  /api/v1/moderation/sessions/{sessionId}/questions:
    get:
      summary: List the new, repeating, priority or complete question queue for one exact tenant session
      description: Repetition is grouped by a server-derived normalized fingerprint; participant text never determines tenant scope.
      security:
        - userSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/WebinarSessionId'
        - name: queue
          in: query
          schema: { type: string, enum: [all, new, repeating, priority], default: new }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/QuestionModerationStatus' }
      responses:
        '200':
          {
            description: Questions with durable status,
            priority,
            revision,
            repeat count and latest human-reviewable suggestion,
          }
        '403': { description: Active membership is not OWNER or MODERATOR }
        '404': { description: Unknown and cross-tenant sessions are intentionally indistinguishable }
  /api/v1/moderation/sessions/{sessionId}/questions/{questionId}:
    patch:
      summary: Change question status or priority with a mandatory reason and optimistic revision
      description: The change writes question history, CRM contact timeline and tenant audit in one transaction.
      security:
        - userSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/WebinarSessionId'
        - $ref: '#/components/parameters/QuestionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuestionModerationCommand' }
      responses:
        '200': { description: Question state, participant history, CRM timeline and audit are synchronized }
        '400': { description: Strict payload or reason validation failed; tenant fields are rejected }
        '403': { description: Active membership is not OWNER or MODERATOR }
        '404': { description: Unknown and cross-tenant questions are intentionally indistinguishable }
        '409': { description: Revision conflict or requested state is already current }
  /api/v1/moderation/sessions/{sessionId}/questions/{questionId}/suggestions:
    post:
      summary: Prepare an unpublished grounded AI-moderator draft
      description: >-
        Searches only the latest published transcript and explicit WebinarSource records. Personalized legal advice
        and questions without a sufficient basis are routed to a human. This operation never publishes a chat message.
      security:
        - userSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/WebinarSessionId'
        - $ref: '#/components/parameters/QuestionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuestionSuggestionGenerateCommand' }
      responses:
        '201': { description: Idempotent pending draft with policy outcome and safe timestamp or HTTPS source citation }
        '400': { description: Strict payload validation failed; tenant fields are rejected }
        '403': { description: Active membership is not OWNER or MODERATOR }
        '404': { description: Unknown and cross-tenant questions are intentionally indistinguishable }
        '409': { description: Question revision is stale or the question is in a terminal state }
  /api/v1/moderation/sessions/{sessionId}/questions/{questionId}/suggestions/{suggestionId}/review:
    post:
      summary: Publish or reject one AI-moderator draft by explicit human action
      description: Publication creates a visibly synthetic AI-moderator message and preserves its safe grounding projection.
      security:
        - userSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/WebinarSessionId'
        - $ref: '#/components/parameters/QuestionId'
        - { name: suggestionId, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 191 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuestionSuggestionReviewCommand' }
      responses:
        '200':
          {
            description: Human decision,
            question state,
            CRM timeline,
            participant history and audit committed atomically,
          }
        '400': { description: Strict payload or reason validation failed; tenant fields are rejected }
        '403': { description: Active membership is not OWNER or MODERATOR }
        '404': { description: Unknown and cross-tenant suggestions are intentionally indistinguishable }
        '409': { description: Question revision is stale or suggestion was already reviewed }
  /api/v1/moderation/sessions/{sessionId}/registrations/{registrationId}/chat-access:
    patch:
      summary: Block or restore chat writes for one exact-session registration
      security:
        - userSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/WebinarSessionId'
        - $ref: '#/components/parameters/RegistrationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegistrationChatAccessCommand' }
      responses:
        '200':
          description: Registration chat state changed, room cache invalidated and audit written
        '400':
          description: Strict payload or reason validation failed; tenant fields are rejected
        '403':
          description: Active membership is not OWNER or MODERATOR
        '404':
          description: Unknown and cross-tenant registrations are intentionally indistinguishable
        '409':
          description: Requested registration chat state is already current
...
  /api/v1/moderation/corrections:
    get:
      summary: List correction requests in the server-selected tenant
      security: [{ userSessionCookie: [] }]
      responses:
        '200':
          {
            description: OWNER sees tenant requests; AUTHOR sees only authored Webinar requests and latest revision status,
          }
        '403': { description: Active membership is not OWNER or AUTHOR }
  /api/v1/moderation/corrections/{requestId}/submissions:
    post:
      summary: Submit an immutable Webinar correction revision for human review
      security: [{ userSessionCookie: [] }]
      parameters:
        - { in: path, name: requestId, required: true, schema: { type: string, minLength: 1, maxLength: 191 } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CorrectionSubmissionRequest' } } }
      responses:
        '201': { description: Private submitted revision; current public Webinar content is unchanged }
        '403': { description: Active membership is not OWNER or AUTHOR }
        '404': { description: Unknown and cross-tenant request are intentionally indistinguishable }
        '409': { description: Request, content version or optimistic revision changed }
...
components:
  parameters:
...
    WebinarSessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 191
...
    ChatMessageId:
      name: messageId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 191 }
    QuestionId:
      name: questionId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 191 }
    RegistrationId:
      name: registrationId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 191 }
  schemas:
...
    WebinarCorrectionContent:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title: { type: string, minLength: 3, maxLength: 180 }
        description: { type: string, nullable: true, minLength: 10, maxLength: 4000 }
        outcomeDescription: { type: string, nullable: true, maxLength: 2000 }
        targetAudience: { type: string, nullable: true, maxLength: 1000 }
        disclaimer: { type: string, nullable: true, maxLength: 2000 }
    CorrectionSubmissionRequest:
      type: object
      additionalProperties: false
      required: [expectedRevision, baseContentVersion, content]
      properties:
        expectedRevision: { type: integer, minimum: 0 }
        baseContentVersion: { type: integer, minimum: 1 }
        content: { $ref: '#/components/schemas/WebinarCorrectionContent' }
...
    ChatMessageModerationCommand:
      type: object
      additionalProperties: false
      required: [action, reason, expectedRevision]
      properties:
        action: { type: string, enum: [HIDE, RESTORE] }
        reason: { type: string, minLength: 3, maxLength: 500 }
        expectedRevision: { type: integer, minimum: 0 }
    QuestionModerationStatus:
      type: string
      enum: [NEW, IN_REVIEW, ACTION_REQUIRED, RESOLVED, REJECTED]
    QuestionPriority:
      type: string
      enum: [NORMAL, HIGH]
    QuestionModerationCommand:
      type: object
      additionalProperties: false
      required: [reason, expectedRevision]
      anyOf:
        - required: [status]
        - required: [priority]
      properties:
        status: { $ref: '#/components/schemas/QuestionModerationStatus' }
        priority: { $ref: '#/components/schemas/QuestionPriority' }
        reason: { type: string, minLength: 3, maxLength: 500 }
        expectedRevision: { type: integer, minimum: 0 }
    QuestionSuggestionGenerateCommand:
      type: object
      additionalProperties: false
      required: [expectedRevision]
      properties:
        expectedRevision: { type: integer, minimum: 0 }
    QuestionSuggestionReviewCommand:
      type: object
      additionalProperties: false
      required: [action, reason, expectedQuestionRevision]
      properties:
        action: { type: string, enum: [PUBLISH, REJECT] }
        reason: { type: string, minLength: 3, maxLength: 500 }
        expectedQuestionRevision: { type: integer, minimum: 0 }
    RegistrationChatAccessCommand:
      type: object
      additionalProperties: false
      required: [action, reason]
      properties:
        action: { type: string, enum: [BLOCK, RESTORE] }
        reason: { type: string, minLength: 3, maxLength: 500 }
...
  securitySchemes:
    metricsBearer:
      type: http
      scheme: bearer
    userSessionCookie:
      type: apiKey
      in: cookie
      name: aspb_user_session
    participantSessionCookie:
      type: apiKey
      in: cookie
      name: aspb_room_token
    adminSessionCookie:
      type: apiKey
      in: cookie
      name: aspb_admin_session